+++ app/libraries/Jentil/Setups/Scripts/CustomizePreview.php
@@ -15,7 +15,7 @@ final class CustomizePreview extends AbstractScript
     {
         parent::__construct($jentil);
 
-        $this->id = "{$this->app->theme->stylesheet}-customize-preview";
+        $this->id = "{$this->app->meta['slug']}-customize-preview";
     }
 
     public function run()