+++ app/libraries/Jentil/Utilities/ThemeMods/Footer.php
@@ -68,7 +68,7 @@ class Footer extends AbstractThemeMod
 
     private function colophonDefault(): string
     {
-        $theme = $this->themeMods->utilities->app->theme;
+        $meta = $this->themeMods->utilities->app->meta;
 
         return \sprintf(
             \esc_html__('&copy; %1$s %2$s. Built with %3$s on %4$s', 'jentil'),
@@ -76,7 +76,7 @@ class Footer extends AbstractThemeMod
             '<a class="blog-name" itemprop="url" href="{{site_url}}">'.
                 '<span itemprop="copyrightHolder">{{site_name}}</span></a>',
             '<em><a itemprop="url" rel="nofollow" href="'.
-                $theme->get('ThemeURI').'">'.$theme->get('Name').'</a></em>',
+                $meta['theme_uri'].'">'.$meta['name'].'</a></em>',
             '<a itemprop="url" rel="nofollow" href="https://wordpress.org">WordPress</a>.'
         );
     }