+++ tests/unit/libraries/Jentil/Setups/MetaBoxes/LayoutTest.php
@@ -15,9 +15,7 @@ class LayoutTest extends AbstractTestCase
         $add_action = FunctionMocker::replace('add_action');
 
         $layout = new Layout(Stub::makeEmpty(AbstractTheme::class, [
-            'theme' => new class {
-                public $stylesheet;
-            },
+            'meta' => ['slug' => 'jentil'],
         ]));
 
         $layout->run();