+++ tests/unit/libraries/Jentil/Setups/Styles/GutenbergTest.php
@@ -17,9 +17,7 @@ class GutenbergTest extends AbstractTestCase
         $add_action = FunctionMocker::replace('add_action');
 
         $style = new Posts(Stub::makeEmpty(AbstractTheme::class, [
-            'theme' => new class {
-                public $stylesheet;
-            },
+            'meta' => ['slug' => 'jentil'],
         ]));
 
         $style->run();
@@ -44,9 +42,7 @@ class GutenbergTest extends AbstractTestCase
 
         $jentil = Stub::makeEmpty(AbstractTheme::class, [
             'utilities' => Stub::makeEmpty(Utilities::class),
-            'theme' => new class {
-                public $stylesheet;
-            },
+            'meta' => ['slug' => 'jentil'],
         ]);
 
         $jentil->utilities->fileSystem = Stub::makeEmpty(FileSystem::class, [