+++ tests/unit/libraries/Jentil/Setups/Thumbnails/MiniTest.php
@@ -15,9 +15,7 @@ class MiniTest extends AbstractTestCase
         $add_action = FunctionMocker::replace('add_action');
 
         $thumbnail = new Mini(Stub::makeEmpty(AbstractTheme::class, [
-            'theme' => new class {
-                public $stylesheet;
-            },
+            'meta' => ['slug' => 'jentil'],
         ]));
 
         $thumbnail->run();
@@ -35,9 +33,7 @@ class MiniTest extends AbstractTestCase
         $add_image_size = FunctionMocker::replace('add_image_size');
 
         $thumbnail = new Mini(Stub::makeEmpty(AbstractTheme::class, [
-            'theme' => new class {
-                public $stylesheet;
-            },
+            'meta' => ['slug' => 'jentil'],
         ]));
 
         $thumbnail->addSize();